WhisperServer/requirements.txt
# Whisper Server Requirements
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
Werkzeug==2.3.7
